// src/app/interfaces/candidaturaObj.ts
export interface ICandidatura {
	name: string
	address: string
	postalCode: string
	local: string
	phone: string
	email: string
	subject: string
	message: string
	candidatura: string
	situacaoActual: string
	ondetrabalha: string
	schedule: string
	literarias: string
	expProfServLimpeza: string
	expProfServOutros: string
}